ヘルパーメソッドをリファクタリング
